pyproject.toml
[tool.poetry]
name = "leadmanagementsystemproject"
version = "0.1.0"
description = ""
authors = ["Madhav Mishra <devopsyodiso@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
django-crispy-forms = "^1.11.1"
django-widget-tweaks = "^1.4.8"
django-simple-history = "^2.12.0"
django-colorful = "^1.3"
django-countries = "^7.1"
django-filter = "^2.4.0"
psycopg2 = "^2.8.6"
requests = "^2.25.1"
Pillow = "^8.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
